requirements.txt
pandas
numpy
scikit-learn
scipy
jieba

torch
transformer
datasets
acceletate
tqdm

statsmodels